refactor and clean up codebase
