+++ src/middleware/validations.js
@@ -1,5 +1,6 @@
 /* eslint-disable linebreak-style */
 import Joi from '@hapi/joi';
+import response from '../response';
 
 const firstName = Joi.string().min(3).max(15).required();
 const lastName = Joi.string().min(3).max(15).required();
@@ -10,7 +11,7 @@ const phoneNumber = Joi.string().regex(/^[0]\d{10}$/).required();
 const productId = Joi.number().integer().required();
 
 class Validation {
-  static signUpValidation(validationObject) {
+  static signUpValidation(req, res, next) {
     const schema = {
       firstName,
       lastName,
@@ -19,22 +20,28 @@ class Validation {
       address,
       phoneNumber,
     };
-    return Joi.validate(validationObject, schema);
+    const { error } = Joi.validate(req.body, schema);
+    if (error) return response(res, 400, error);
+    return next();
   }
 
-  static loginValidation(validationObject) {
+  static loginValidation(req, res, next) {
     const schema = {
       email,
       password,
     };
-    return Joi.validate(validationObject, schema);
+    const { error } = Joi.validate(req.body, schema);
+    if (error) return response(res, 400, error);
+    return next();
   }
 
-  static addToCart(validationObject) {
+  static addToCart(req, res, next) {
     const schema = {
       productId,
     };
-    return Joi.validate(validationObject, schema);
+    const { error } = Joi.validate(req.params, schema);
+    if (error) return response(res, 400, error);
+    return next();
   }
 }
