+++ src/routes/cart.js
@@ -2,10 +2,11 @@
 import express from 'express';
 import Authentication from '../Authentication';
 import CartController from '../controllers/Cart';
+import Validation from '../middleware/validations';
 
 const router = express.Router();
 
-router.post('/:productId', Authentication.verifyToken, CartController.addOrDeductFromCart);
+router.post('/:productId', Authentication.verifyToken, Validation.addToCart, CartController.addOrDeductFromCart);
 
 router.get('/', Authentication.verifyToken, CartController.getCart);
