+++ src/routes/user.js
@@ -1,12 +1,13 @@
 /* eslint-disable linebreak-style */
 import express from 'express';
+import Validation from '../middleware/validations';
 
 import UserController from '../controllers/User';
 
 const router = express.Router();
 
-router.post('/signup', UserController.signup);
+router.post('/signup', Validation.signUpValidation, UserController.signup);
 
-router.post('/signin', UserController.signin);
+router.post('/signin', Validation.loginValidation, UserController.signin);
 
 export default router;
